[add, update] ファイルの追加とモデルに関する箇所の移動
